TTSAlbion/Python/requirements.txt
discord.py[voice]>=2.4,<3.0
